Improved task

+++ src/main/kotlin/g2801_2900/s2875_minimum_size_subarray_in_infinite_array/Solution.kt
@@ -1,7 +1,7 @@
 package g2801_2900.s2875_minimum_size_subarray_in_infinite_array
 
 // #Medium #Array #Hash_Table #Prefix_Sum #Sliding_Window
-// #2023_12_25_Time_388_ms_(100.00%)_Space_50.3_MB_(100.00%)
+// #2023_12_25_Time_372_ms_(100.00%)_Space_52.6_MB_(100.00%)
 
 import kotlin.math.min
 
@@ -26,7 +26,7 @@ class Solution {
                 start++
             }
             if (currentSum == sum) {
-                min = min(min.toDouble(), (i - start + 1).toDouble()).toInt()
+                min = min(min, i - start + 1)
             }
         }
         if (min == nums.size) {
